# ORC-1170:[C++] Optimize the RowReader::seekToRow function in the case of seeking within the same stripe