### What changes were proposed in this pull request?
The patch will optimize the seekToRow function. The RowReader::seekToRow() function always call startNextStripe() to seek to the beginning of target stripe. However, when PPD is enabled many calls are still staying in the same stripe which results in many unnecessary calls. 

### Why are the changes needed?
This PR improves the seekToRow function call. I pick a 259MB file of the TPCH lineitem table which contains 7200030
records. I use a modified orc-scan tool that calls the RowReader::seekToRow function once before each call the RowReader:: next() function to test this patch.  The results are shown below:

|| original | optimized | speedup |
|:-----|:----:|:----:|:----:|
|Elapsed Time(s)|43.69|27.13|1.61|


### How was this patch tested?
The original UTs(TestPredicatePushdown, DictionaryEncoding::multipleStripesWithIndex, DictionaryEncoding::multipleStripesWithoutIndex) will cover this patch.