@coderex2522 2022-05-16
> LGTM. It'd be nice if we can have a more specifit commit tittle. We are optimzing the case seeking within the same stripe. Seeking within the same RowGroup is not optimized in this PR. So having a title like "[ORC-1170](https://issues.apache.org/jira/browse/ORC-1170):[C++] Optimize seeking within the same stripe" seems better to me.

I think seeking within the same stripe  is inclusive of  seeking withing the same RowGroup. Seeking withing the same RowGroup will also call function startNextStripe under the original implementation.