@dongjoon-hyun 2022-05-21
This is merged via https://github.com/apache/orc/commit/11a2544f4e54b29d5b0d3ca6aeb8dde5291b2163